Split the CI workflows into the release and non-release.
This helps to separate between the two and not perform the steps we
don't need all the time.

The ci-basic.yml contains only the steps which do not involve the
release steps, such as publishing the artifacts to the S3 storage and
uploading the docker image to the registry. It also contains fewer
jobs, which are the most common ones: the recent ubuntu versions,
valgrind and address sanitizer.

The ci-full.yml contains all the steps for all the supported platforms,
and involves publishing the artifacts and the docker image. This
workflow automatically determines the S3 destination based on the
branch name and the tag name: if the push or the pull request is made
with a release tag, it is considered a production build, otherwise -
a staging build, and the S3 destination is chosen accordingly.